% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsNB.R
\name{controlNB}
\alias{controlNB}
\title{Define controllable parameters of initial NB fit}
\usage{
controlNB(
  start = list(mu = NULL, logTheta = NULL),
  method = "BFGS",
  controlOptim = list(maxit = 100),
  initThetaMASS = FALSE,
  initMASS = TRUE,
  restricted = FALSE,
  eps = .Machine$double.eps^0.25,
  epsilonMASS = 1e-08
)
}
\arguments{
\item{start}{Optional starting values for \link[WALS]{fitNB2}. Only used if
\code{initMASS = FALSE}.}

\item{method}{Optimization method used in \link[stats]{optim}. Only used if
\code{initMASS = FALSE}.}

\item{controlOptim}{List with parameters controlling optimization process of
\link[stats]{optim}. Only used if \code{initMASS = FALSE}.}

\item{initThetaMASS}{If TRUE, then initial \eqn{\log{\theta}} of
\link[WALS]{fitNB2} is estimated using \link[MASS]{theta.ml}
(ML-estimation over 1 variable) based on regression coefficients from
Poisson regression. If \code{FALSE}, then initial \eqn{\log{theta}} = 0 is used.}

\item{initMASS}{If TRUE (default), then initial fit in \link[WALS]{fitNB2} is estimated via
\link[MASS]{glm.nb} and \code{initThetaMASS} is ignored.}

\item{restricted}{If TRUE, then initial fit in \link[WALS]{fitNB2} only considers the
focus regressors. By default \code{FALSE}, then the unrestricted model is
estimated in \link[WALS]{fitNB2} (i.e. all regressors).}

\item{eps}{Controls argument \code{eps} in \link[WALS]{fitNB2} for generating
starting value for \code{logTheta} (\eqn{\log{\theta}}) via \link[MASS]{theta.ml}.}

\item{epsilonMASS}{Sets epsilon in control argument of \link[MASS]{glm.nb}.}
}
\description{
Define controllable parameters of initial NB fit
}
