% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{ddgengamma}
\alias{ddgengamma}
\title{Double generalized gamma density}
\usage{
ddgengamma(x, q, alpha, b, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{q}{\eqn{q} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}

\item{alpha}{\eqn{\alpha} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Wrapper around dgengamma.stacy() of VGAM to use the parametrization on pp. 131
of \insertCite{magnus2016wals;textual}{WALS}.
}
\details{
The density function is
\deqn{\pi(x) = \frac{q c^{(1 - \alpha)/q}}{2 \Gamma((1 - \alpha)/q)} |x|^{-\alpha} \exp(-c |x|^{q})}.

The function uses \code{dgengamma.stacy} internally from \code{VGAM} that uses
the parametrization in table 12.13, p.369 of \insertCite{yee2015vgam;textual}{WALS}.
}
\references{
\insertAllCited{}
}
