% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{ddsubbotin}
\alias{ddsubbotin}
\title{Subbotin density}
\usage{
ddsubbotin(x, q, b, log = FALSE)
}
\arguments{
\item{q}{\eqn{q} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Wrapper around dgengamma.stacy() of VGAM to use the parametrization on pp. 131
of \insertCite{magnus2016wals;textual}{WALS}.
}
\references{
\insertAllCited{}
}
