% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{dlaplace}
\alias{dlaplace}
\title{Laplace density}
\usage{
dlaplace(x, b, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\description{
Wrapper around dgengamma.stacy() of VGAM to use the parametrization on pp. 131
of \insertCite{magnus2016wals;textual}{WALS}.
}
\references{
\insertAllCited{}
}
