% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{familyPrior}
\alias{familyPrior}
\alias{weibull}
\alias{subbotin}
\alias{laplace}
\alias{familyPrior_laplace}
\title{Family Objects for Prior Distributions in WALS}
\usage{
familyPrior(object, ...)

weibull(q = 0.887630085544086, b = log(2))

subbotin(q = 0.799512530172489, b = 0.937673273794677)

laplace(b = log(2))
}
\arguments{
\item{object}{Object of of class \code{familyPrior}.}

\item{...}{Further arguments passed to methods.}

\item{q}{\eqn{q} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution (\link{ddgengamma}).}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution (\link{ddgengamma}).}
}
\value{
An object of class \code{familyPrior}. This is a list with the
elements
\item{q}{Parameter \eqn{q}.}
\item{alpha}{Parameter \eqn{\alpha} (of the reflected generalized gamma
distribution).}
\item{b}{Parameter \eqn{c}.}
\item{delta}{Parameter \eqn{\delta = (1 - \alpha)/q}.}
\item{printPars}{vector. Contains the parameters that are shown in printing
functions, e.g. \code{print.familyPrior}.}.
\item{prior}{String with the name of the prior distribution.}

\code{laplace} returns an object of the specialized class
\code{familyPrior_laplace} that inherits from \code{familyPrior}.
This allows separate processing of the Laplace prior in the estimation
functions as closed-form formulas exists for its posterior mean and variance.
The list elements are the same as for objects of class \code{familyPrior}.
}
\description{
\code{familyPrior} objects provide a convenient way to specify the prior
distribution usedfor the Bayesian posterior mean estimation of the WALS
estimators in \link{wals}, \link{walsGLM} and \link{walsNB}.
}
\details{
The double (reflected) Weibull \link{weibull}, Subbotin \link{subbotin} and
Laplace distributions (\link{laplace}) are all special cases of the reflected
generalized gamma distribution (\link{ddgengamma}). The Laplace distribution
is also a special case of the double Weibull and of the Subbotin distribution.

The default values for the parameters \code{q} and \code{b} are minimax regret
solutions for the corresponding priors. The double (reflected) Weibull and
Subbotin prior are both neutral and robust. In contrast, the Laplace prior
is only neutral but not robust. See section 9 "Enter Bayes: Neutrality and
Robustness" of \insertCite{magnus2016wals;textual}{WALS} for details.
}
\references{
\insertAllCited{}
}
\seealso{
\link{wals}, \link{walsGLM}, \link{walsNB}, \link{computePosterior}, \link{ddweibull},
\link{dsubbotin}, \link{dlaplace}, \link{ddgengamma}.
}
