% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyWALS.R
\name{familyWALS}
\alias{familyWALS}
\alias{poissonWALS}
\alias{binomialWALS}
\alias{negbinFixedWALS}
\alias{negbinWALS}
\alias{familyNBWALS}
\title{Extended Family Objects for Models}
\usage{
familyWALS(object, ...)

poissonWALS(link = "log")

binomialWALS(link = "logit")

negbinFixedWALS(scale, link)

negbinWALS(scale, link)
}
\arguments{
\item{object}{Object of of class \code{familyWALS}.}

\item{...}{Further arguments passed to methods.

The \code{negbinWALS} family currently only accepts \code{"log"}, while
\code{negbinFixedWALS} supports both \code{"log"} and \code{"canonical"}.}

\item{link}{Specifies the model link function. Typically a character string
or an object of class \code{link-glm} generated by \link[stats]{make.link}.
See \link[stats]{family} for more details.}

\item{scale}{dispersion parameter of NB2 to be used, always larger than 0.}
}
\value{
An object of class \code{familyWALS} that inherits from
\link[stats]{family}. This is a list that contains elements returned from the
corresponding family function that it extends. Additionally, the following
elements are available:
\item{theta.eta}{function. Derivative of the canonical parameter \eqn{\theta}
with respect to the linear link \eqn{\eta}, i.e. \eqn{d \theta / d \eta}.}
\item{psi}{function. \eqn{\psi} defined on p. 3 of \insertCite{deluca2018glm}{WALS}.}
\item{initializeY}{function. Preprocesses the response, e.g. in binomialWALS()
it transforms factors to numeric 0s and 1s.}
\item{transformY}{function. Transforms the response to \eqn{\bar{y}}.
See eq. (5) in \insertCite{deluca2018glm}{WALS} for GLMs and
\insertCite{huynhwalsnb}{WALS} for \code{negbinWALS} used in \link{walsNB}.}
\item{transformX}{function. Transforms the regressors to \eqn{\bar{X}_1} and
\eqn{\bar{X}_2}, respectively. See eq. (5) in \insertCite{deluca2018glm}{WALS}
for GLMs and \insertCite{huynhwalsnb}{WALS} for \code{negbinWALS} used in \link{walsNB}.}
\item{density}{function. The probability density/mass function of the family.}

\code{negbinWALS} creates an object of class \code{familyNBWALS} that
inherits from \code{familyWALS} and \link[stats]{family}. This is a list that
contains all elements returned from the corresponding family function that it
extends and the elements described above for objects of class \code{familyWALS}.
Additionally contains the following elements with functions required in
\link{walsNB} that are described in \insertCite{huynhwalsnb}{WALS}:
\item{q}{function. Computes \eqn{\bar{q}}.}
\item{g}{function. Computes \eqn{\bar{g}}.}
\item{transformY0}{function. Computes \eqn{\bar{y}_0}.}
\item{t}{function. Computes \eqn{\bar{t}}.}
\item{epsilon}{function. Computes \eqn{\bar{\epsilon}}.}
\item{epsiloninv}{function. Computes \eqn{\bar{\epsilon}^{-1}}.}
\item{kappaSum}{function. Computes \eqn{\bar{\kappa}^{\top} \mathbf{1}}.}
\item{computeAlpha}{function. Computes the log-dispersion parameter
\eqn{\log(\rho)} given (model-averaged) estimates of the regression
coefficients of the transformed regressors \eqn{\gamma_{1}} and \eqn{\gamma_{2}}.}
}
\description{
Objects of class \code{familyWALS} inherit from \link[stats]{family} and
extend those with (transformation) functions required for \link[WALS]{walsGLM}
and \link[WALS]{walsNB}.
}
\details{
\code{familyWALS} is a generic function that extracts the family used in
\code{walsGLM} objects.

\code{negbinFixedWALS} is the \code{familyWALS} object for negative binomial
distribution type 2 (NB2) with fixed dispersion parameter. It extends
\link{negativeBinomial}.

\code{negbinWALS} is a specialized object of class \code{familyNBWALS} which
inherits from \code{familyWALS} and \code{family}. It constructs the
\code{familyNBWALS} object for the negative binomial distribution type 2 (NB2)
with variable dispersion parameter and by extending \link{negativeBinomial}
and \link{negbinFixedWALS} with functions required in \link{walsNB}.
\strong{\code{negbinWALS} should only be used in \link{walsNBfit} and not in
\link{walsGLM} because the NB2 with variable dispersion parameter is not a GLM!}
}
\references{
\insertAllCited{}
}
\seealso{
\link{family}.
}
