% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyWALShurdle.R
\name{familyWALShurdle}
\alias{familyWALShurdle}
\title{Constructor of familyWALShurdle}
\usage{
familyWALShurdle(count, zero)
}
\description{
Uses a zerotruncated family and a family for the zero-part to construct
a joint density function used in predict.walsHurdle
}
