% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNB2.R
\name{fitNB2}
\alias{fitNB2}
\title{Fits a NB2 regression via maximum likelihood with log link for mean and
dispersion parameter.}
\usage{
fitNB2(X, Y, family, control = controlNB())
}
\arguments{
\item{X}{Design matrix.}

\item{Y}{Count response vector.}

\item{family}{Object of class \code{\link[stats]{family}}.}

\item{control}{List of parameters for controlling the optimization process.
Use \link[WALS]{controlNB} to generate the list.}
}
\description{
Only works with log-link so far, no other links tested.
}
\details{
The available parameters for controlling the optimization are documented in
\link[WALS]{controlNB}.
}
