% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{gammaToBeta}
\alias{gammaToBeta}
\title{Transform gammas back to betas}
\usage{
gammaToBeta(
  posterior,
  y,
  Z1,
  Z2,
  Delta1,
  D2,
  sigma,
  Z1inv,
  method = "original",
  svdZ1
)
}
\description{
Transforms posterior means (gammas) and variances of transformed Z2
back to regression coefficients (betas) of original regressors X1 and X2.
}
