% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{laplace}
\alias{laplace}
\title{Laplace prior}
\usage{
laplace(b = log(2))
}
\arguments{
\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}
}
\description{
Laplace prior for Bayesian posterior mean estimation of WALS estimators.
Special case of the reflected generalized gamma distribution and of
the Weibull and Subbotin prior.
}
\details{
The default value for the parameter \code{b} is the minimax regret
solution. The Laplace prior is only neutral but not robust.
See section 9 "Enter Bayes: Neutrality and Robustness" of
\insertCite{magnus2016wals;textual}{WALS} for details.
}
\references{
\insertAllCited{}
}
