% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyWALS.R
\name{negativeBinomial}
\alias{negativeBinomial}
\title{Negative binomial family}
\usage{
negativeBinomial(theta, link = "log")
}
\arguments{
\item{theta}{dispersion parameter of NB2, always larger than 0.}

\item{link}{specifies link function, currently only "log" and "canonical"
are supported.}
}
\description{
Reconstruct family object for negative binomial type 2 (NB2) with fixed
scale parameter theta. Analogous to \code{\link[MASS]{negative.binomial}} in
\code{MASS} \insertCite{mass2002}{WALS} but \code{MASS} uses non-canonical link.
}
\references{
\insertAllCited{}
}
\seealso{
\link{family}, \link{familyWALS}, \link{negbinWALS}, \link{negbinFixedWALS}.
}
