% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyWALS.R
\name{negbinFixedWALS}
\alias{negbinFixedWALS}
\title{NB2 family with fixed dispersion parameter and additional functions}
\usage{
negbinFixedWALS(scale, link)
}
\arguments{
\item{scale}{dispersion parameter of NB2, always larger than 0.}

\item{link}{specifies link function, currently only "log" and "canonical"
are supported.}
}
\description{
Family object for negative binomial distribution type 2 (NB2) with fixed
dispersion parameter. Extends \link{negativeBinomial} with functions
required in \code{walsGLM}.
}
