% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyWALS.R
\name{negbinWALS}
\alias{negbinWALS}
\title{NB2 family with variable scale parameter and additional functions}
\usage{
negbinWALS(scale, link)
}
\arguments{
\item{scale}{dispersion parameter of NB2 to be used.}

\item{link}{specifies the link function, currently only "log" is supported.}
}
\description{
Family object for negative binomial distribution type 2 (NB2) with variable
dispersion parameter. Extends \link{negativeBinomial} with functions
required in \link{walsNB}.
}
