% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyWALS.R
\name{poissonWALS}
\alias{poissonWALS}
\title{Poisson family with additional functions}
\usage{
poissonWALS(link = "log")
}
\arguments{
\item{link}{specifies link for \code{poisson} family.
See \link[stats]{family} for details.}
}
\description{
Extends \code{poisson} family from \code{stats} with transformations required
for \code{walsGLM}.
}
