% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wals.R
\name{predict.wals}
\alias{predict.wals}
\alias{predict.walsMatrix}
\alias{fitted.wals}
\alias{residuals.wals}
\alias{print.wals}
\alias{summary.wals}
\alias{print.summary.wals}
\alias{coef.wals}
\alias{vcov.wals}
\alias{nobs.wals}
\alias{terms.wals}
\alias{model.matrix.wals}
\title{Methods for wals and walsMatrix Objects}
\usage{
\method{predict}{wals}(object, newdata, na.action = na.pass, ...)

\method{predict}{walsMatrix}(object, newX1, newX2, ...)

\method{fitted}{wals}(object, ...)

\method{residuals}{wals}(object, ...)

\method{print}{wals}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{wals}(object, ...)

\method{print}{summary.wals}(x, digits = max(3, getOption("digits") - 3), ...)

\method{coef}{wals}(object, type = c("all", "focus", "aux"), transformed = FALSE, ...)

\method{vcov}{wals}(object, type = c("all", "focus", "aux"), transformed = FALSE, ...)

\method{nobs}{wals}(object, ...)

\method{terms}{wals}(x, type = c("focus", "aux"), ...)

\method{model.matrix}{wals}(object, type = c("focus", "aux"), ...)
}
\arguments{
\item{object, x}{An object of class \code{wals} or \code{walsMatrix}.}

\item{newdata}{Optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{na.action}{Function determining what should be done with missing values
in \code{newdata}. The default is to predict \code{NA}.}

\item{...}{Further arguments passed to methods.}

\item{newX1}{Focus regressors matrix to be used for the prediction.}

\item{newX2}{Auxiliary regressors matrix to be used for the prediction.}

\item{digits}{The number of significant digits to show.}

\item{type}{Character specifying the part of the model that should be returned.
For details see below.}

\item{transformed}{Logical specifying whether the coefficients/covariance
matrix of original regressors (\code{FALSE}, default) or the transformed
regressors (\code{TRUE}) should be returned.}
}
\description{
Methods for extracting information from fitted model-averaging objects of
classes \code{wals} and \code{walsMatrix}. \code{walsMatrix} objects inherit
from \code{wals}, so the methods for \code{wals} also work for objects of
class \code{walsMatrix}.
}
\details{
A set of standard extractor functions for fitted model objects is available
for objects of class \code{wals} and \code{walsMatrix}, including methods to
the generic functions \code{\link[base]{print}} and \code{\link[base]{summary}}
which print the model-averaged estimation of the coefficients along with some
further information. As usual, the \code{summary} method returns an object of
class \code{"summary.wals"} containing the relevant summary statistics which
can then be printed using the associated \code{print} method.
Inspired by \insertCite{deluca2011stata;textual}{WALS},the summary statistics
also show \code{Kappa} which is an indicator for the numerical stability of
the method, i.e. it shows the square root of the condition number of the
matrix \eqn{\Xi = \Delta_{2} X_{2}^{\top} M_{1} X_{2} \Delta_{2}}.
The summary further provides information on the prior used along with its
parameters.

For \code{\link[stats]{coef}} and \code{\link[stats]{vcov}}, the \code{type}
argument specifies which part of the coefficient vector/covariance matrix of
the estimates should be returned. For \code{type = "all"}, they return the
complete vector/matrix. For \code{type = "focus"} and \code{type = "aux"} they
return only the part corresponding to the focus and auxiliary regressors,
respectively. Additionally, the user can choose whether to return the
estimated coefficients/covariance matrix for the original regressors \eqn{X}
(\eqn{\beta} coefficients) or of the transformed regressors \eqn{Z}
(\eqn{\gamma} coefficients).

The extractors \code{\link[stats]{terms}} and \code{\link[stats]{model.matrix}}
behave similarly to \code{coef.wals}, but they only allow \code{type = "focus"}
and \code{type = "aux"}. They extract the corresponding component of the model.

The \code{\link[stats]{residuals}} method computes raw residuals
(observed - fitted).

The \code{\link[WALS]{familyPrior}} method returns an object of class
\code{familyPrior} that was used as prior in the Bayesian estimation step.
}
