% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsGLM.R
\name{predict.walsGLM}
\alias{predict.walsGLM}
\alias{predict.walsGLMmatrix}
\alias{residuals.walsGLM}
\alias{print.walsGLM}
\alias{summary.walsGLM}
\alias{print.summary.walsGLM}
\alias{logLik.walsGLM}
\alias{familyWALS.walsGLM}
\title{Methods for walsGLM and walsGLMmatrix Objects}
\usage{
\method{predict}{walsGLM}(
  object,
  newdata,
  type = c("response", "link", "variance", "prob", "density", "logDens"),
  at = NULL,
  na.action = na.pass,
  log = FALSE,
  ...
)

\method{predict}{walsGLMmatrix}(
  object,
  newX1,
  newX2,
  newY = NULL,
  type = c("response", "link", "variance", "prob", "density", "logDens"),
  at = NULL,
  log = FALSE,
  ...
)

\method{residuals}{walsGLM}(object, type = c("deviance", "pearson", "response"), ...)

\method{print}{walsGLM}(x, digits = max(3, getOption("digits") - 3), ...)

\method{summary}{walsGLM}(object, ...)

\method{print}{summary.walsGLM}(x, digits = max(3, getOption("digits") - 3), ...)

\method{logLik}{walsGLM}(object, ...)

\method{familyWALS}{walsGLM}(object, ...)
}
\arguments{
\item{object, x}{An object of class \code{wals} or \code{walsMatrix}.}

\item{newdata}{Optionally, a data frame in which to look for variables with
which to predict. If omitted, the original observations are used.}

\item{type}{Character specifying the type of prediction, residual or model
part to be returned. For details see below.}

\item{at}{Optional. Only available if a family of class \code{familyWALScount}
was used for fitting. If \code{type = "prob"}, a numeric vector at which
the probabilities are evaluated. By default \code{0:max(y)} is used
where \code{y} is the original observed response.}

\item{na.action}{Function determining what should be done with missing values
in \code{newdata}. The default is to predict \code{NA}.}

\item{log}{Logical. If \code{TRUE}, then returns the log-density. If
\code{FALSE} (default), then returns density. Only relevant if
\code{type = "density"}.}

\item{...}{Further arguments passed to methods.}

\item{newX1}{Focus regressors matrix to be used for the prediction.}

\item{newX2}{Auxiliary regressors matrix to be used for the prediction.}

\item{newY}{Response vector to be used in predictions. Only relevant when
\code{type = "prob"}, \code{type = "density"} or \code{type = "logDens"}.}

\item{digits}{The number of significant digits to show.}
}
\description{
Methods for extracting information from fitted model-averaging objects of
classes \code{walsGLM} and \code{walsGLMmatrix}. \code{walsGLMmatrix} objects
inherit from \code{walsGLM}, so the methods for \code{walsGLM} also work for
objects of class \code{walsGLMmatrix}.
}
\details{
A set of standard extractor functions for fitted model objects is available
for objects of class \code{walsGLM} and \code{walsGLMmatrix}, including methods to
the generic functions \code{\link[base]{print}} and \code{\link[base]{summary}}
which print the model-averaged estimation of the coefficients along with some
further information. As usual, the \code{summary} method returns an object of
class \code{"summary.walsGLM"} containing the relevant summary statistics which
can then be printed using the associated \code{print} method.
Inspired by \insertCite{deluca2011stata;textual}{WALS},the summary statistics
also show \code{Kappa} which is an indicator for the numerical stability of
the method, i.e. it shows the square root of the condition number of the
matrix \eqn{\bar{\Xi} = \bar{\Delta}_{2} \bar{X}_{2}^{\top} \bar{M}_{1}
\bar{X}_{2} \bar{\Delta}_{2}}. The summary further shows the deviance and
provides information on the prior and family used.

A \code{\link[stats]{logLik}} method is provided that returns the log-likelihood
given the family used and the model-averaged estimates of the coefficients.

\code{walsGLM} and \code{walsGLMmatrix} inherit from \code{wals} and
\code{walsMatrix}, respectively. See \link[WALS]{predict.wals} for more
methods.
}
\section{Details on the use of the argument type}{
For \code{\link[stats]{predict}} and \code{\link[stats]{residuals}}, the
\code{type} argument specifies the type of prediction or residual to return.

Different return types for \code{\link[stats]{predict}}:
\itemize{
\item \code{type = "response"}: predicted mean
\item \code{type = "link"}: predicted linear link
\item \code{type = "variance"}: predicted variance
\item \code{type = "prob"}: Only available if a family of class \code{familyWALScount}
was used for fitting. Returns the probability at counts specified by \code{at}.
\item \code{type = "density"}: predicted density
\item \code{type = "logDens"}: for convenience, returns predicted log-density.
Equivalent to setting \code{type = "density"} and \code{log = TRUE}.
If \code{type = "prob"}, \code{type = "density"} or \code{type = "logDens"},
then \code{newdata} must contain the response or \code{newY} must be
specified depending on the class of the object.
}

Different return types for \code{\link[stats]{residuals}}:
\itemize{
\item \code{type = "deviance"}: deviance residuals
\item \code{type = "pearson"}: Pearson residuals (raw residuals scaled by
square root of variance function)
\item \code{type = "response"}: raw residuals (observed - fitted)
}

For \code{\link[stats]{coef}} and \code{\link[stats]{vcov}}, the \code{type}
argument specifies which part of the coefficient vector/covariance matrix of
the estimates should be returned. For \code{type = "all"}, they return the
complete vector/matrix. For \code{type = "focus"} and \code{type = "aux"} they
return only the part corresponding to the focus and auxiliary regressors,
respectively. Additionally, the user can choose whether to return the
estimated coefficients/covariance matrix for the original regressors \eqn{X}
(\eqn{\beta} coefficients) or of the transformed regressors \eqn{Z}
(\eqn{\gamma} coefficients).

The extractors \code{\link[stats]{terms}} and \code{\link[stats]{model.matrix}}
behave similarly to \code{coef}, but they only allow \code{type = "focus"}
and \code{type = "aux"}. They extract the corresponding component of the model.
}

\references{
\insertAllCited{}
}
\seealso{
\link{walsGLM}, \link{predict.wals}
}
