% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsZerotrunc.R
\name{predict.walsZerotrunc}
\alias{predict.walsZerotrunc}
\title{Predict method for walsZerotrunc Fits}
\usage{
\method{predict}{walsZerotrunc}(
  object,
  newdata,
  type = c("response", "link", "variance", "zero", "count", "prob"),
  na.action = na.pass,
  at = NULL,
  ...
)
}
\arguments{
\item{type}{the type of prediction required. The default (\code{type = "response"})
is the mean of the zerotruncated ddistribution, i.e. it is on the scale of the response variable.
\code{type = "link"} gives the linear predictor, \code{type = "variance"}
the variance of the zerotruncated distribution. Further, \code{type = "zero"}
returns the probability for observing counts larger than zero (i.e. P(Y > 0))
and \code{type = "count"} returns the predicted mean of the untruncated distribution.
Finally, \code{type = "prob"} returns the probability for counts.}
}
\description{
Obtains predictions from a fitted walsZerotrunc object
}
