% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitNB2.R
\name{snbinom}
\alias{snbinom}
\title{Internal function: first derivatives of NB2 PMF}
\usage{
snbinom(x, mu, size, parameter = c("mu", "size"), drop = TRUE)
}
\arguments{
\item{x}{Vector of quantiles.}

\item{mu}{Vector of means.}

\item{size}{Vector of dispersion parameter. If a scalar is given, the value
is recycled.}

\item{parameter}{Specifies which parameter the derivative is taken for.
\code{parameter = c("mu", "size")} returns a matrix with derivatives
for both parameters.}

\item{drop}{If \code{TRUE}, drops empty dimensions of return using
\code{\link[base]{drop}}. If \code{FALSE} does not apply \code{\link[base]{drop}}.}
}
\value{
A vector or matrix containing the first derivatives.
}
\description{
First derivatives of NB2 PMF used in \code{\link[WALS]{fitNB2}}. Code is
taken from the function \code{snbinom()} in the \code{countreg} package
version 0.2-1 (2023-06-13) \insertCite{countreg}{WALS}.
}
\references{
\insertAllCited{}
}
