% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{subbotin}
\alias{subbotin}
\title{Subbotin prior}
\usage{
subbotin(q = 0.799512530172489, b = 0.937673273794677)
}
\arguments{
\item{q}{\eqn{q} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}
}
\description{
Subbotin prior for Bayesian posterior mean estimation of WALS estimators.
Special case of the reflected generalized gamma distribution.
}
\details{
The default values for the parameters \code{q} and \code{b} are minimax regret
solutions. The Subbotin prior is both neutral and robust.
See section 9 "Enter Bayes: Neutrality and Robustness" of
\insertCite{magnus2016wals;textual}{WALS} for details.
}
\references{
\insertAllCited{}
}
