% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsNB.R
\name{vcov.walsNB}
\alias{vcov.walsNB}
\title{Calculate Variance-Covariance Matrix for a \code{"walsNB"} object}
\usage{
\method{vcov}{walsNB}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"walsNB"}.}

\item{...}{For expansion in the future.}
}
\value{
No return value, only raises error because no covariance matrix
estimator exists yet.
}
\description{
This method always raises an error because the covariance matrix of the
walsNB estimator has not been derived yet.
}
