% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wals.R
\name{wals.fit}
\alias{wals.fit}
\title{Fitter function for Weighted Average Least Squares estimation}
\usage{
wals.fit(
 X1,
 X2,
 y,
 sigma = NULL,
 prior = weibull(),
 method = "original",
 svdTol = .Machine$double.eps,
 svdRtol = 1e-6,
 keepUn = FALSE,
 eigenSVD = TRUE,
 prescale = TRUE,
 ...
 )
}
\arguments{
\item{X1}{Design matrix for focus regressors. Usually includes a constant
(column full of 1's) and can be generated using model.matrix().}

\item{X2}{Design matrix for auxiliary regressors. Usually does not include
a constant column and can also be generated using model.matrix().}

\item{y}{Count response as vector}

\item{sigma}{if NULL (default), then the variance of the error term is estimated,
see p.136 of \insertCite{magnus2016wals;textual}{WALS}. If sigma is specified,
then the unrestricted estimator is divided by sigma before performing the
Bayesian posterior mean estimation.}

\item{prior}{Object of class \code{familyPrior}. For example \link[WALS]{weibull}
or \link[WALS]{laplace}. Not tested with other priors.}

\item{method}{Specifies method used. Available methods are
\code{"original"} (default) or \code{"svd"}.}

\item{svdTol}{Tolerance for rank of matrix \eqn{\bar{Z}_{1}}
Only used if \code{method = "svd"}.
Checks if smallest eigenvalue in SVD of \eqn{\bar{Z}_1} and \eqn{\bar{Z}}
is larger than \code{svdTol}, otherwise reports a rank deficiency.}

\item{svdRtol}{Relative tolerance for rank of matrix \eqn{\bar{Z}_{1}}.
Only used if \code{method = "svd"}. Checks if ratio of largest to smallest
eigenvalue in SVD of \eqn{\bar{Z}_1} is larger than  \code{svdRtol},
otherwise reports a rank deficiency.}

\item{keepUn}{If \code{TRUE}, keeps the estimators of the unrestricted model,
i.e. \eqn{\tilde{\gamma}_{u}}.}

\item{eigenSVD}{If \code{TRUE}, then \code{semiorthogonalize()} uses \code{svd()}
to compute the eigendecomposition of \eqn{\bar{\Xi}} instead of \code{eigen()}.
In this case, the tolerances of \code{svdTol} and \code{svdRtol} are used to
determine whether \eqn{\bar{\Xi}} is of full rank (need it for \eqn{\bar{\Xi}^{-1/2}}).}

\item{prescale}{If \code{TRUE} (default), prescales the regressors X1 and X2 with
\eqn{\Delta_1} and \eqn{\Delta_2}, respectively, to improve numerical stability
and make the coefficients of the auxiliary regressors scale equivariant.
See \insertCite{deluca2011stata;textual}{WALS} for more details.
\strong{WARNING: It is not recommended to set \code{prescale = FALSE}.}
The option \code{prescale = FALSE} only exists for historical reasons.}

\item{...}{Arguments for internal function \code{computePosterior()}.}
}
\description{
Workhorse function behind \link{wals} and \link{walsGLM}.
}
\references{
\insertAllCited{}
}
