% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsGLM.R
\name{walsGLM.fit}
\alias{walsGLM.fit}
\title{Fitter function for Weighted Average Least Squares estimation of GLMs}
\usage{
walsGLM.fit(
 X1,
 X2,
 y,
 betaStart1,
 betaStart2,
 family,
 prior = weibull(),
 ...)
}
\description{
Fitter function for Weighted Average Least Squares estimation of GLMs
}
