% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsGLM.R
\name{walsGLM}
\alias{walsGLM}
\alias{walsGLM.default}
\alias{walsGLM.matrix}
\alias{walsGLM.formula}
\title{Weighted Average Least Squares for Generalized Linear Models}
\usage{
walsGLM(x, ...)

\method{walsGLM}{default}(x, ...)

\method{walsGLM}{matrix}(
  x,
  x2,
  y,
  family,
  subset = NULL,
  na.action = NULL,
  weights = NULL,
  offset = NULL,
  prior = weibull(),
  controlGLMfit = list(),
  keepY = TRUE,
  keepX = FALSE,
  iterate = FALSE,
  tol = 1e-06,
  maxIt = 10000,
  nIt = NULL,
  verbose = FALSE,
  ...
)

\method{walsGLM}{formula}(
  formula,
  family,
  data,
  subset = NULL,
  na.action = NULL,
  weights = NULL,
  offset = NULL,
  prior = weibull(),
  controlGLMfit = list(),
  model = TRUE,
  keepY = TRUE,
  keepX = FALSE,
  iterate = FALSE,
  tol = 1e-06,
  maxIt = 10000,
  nIt = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{...}{Arguments for workhorse \link[WALS]{walsGLMfit}.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{\strong{not implemented yet.}}

\item{weights}{\strong{not implemented yet.}}

\item{offset}{\strong{not implemented yet.}}

\item{iterate}{if TRUE then the WALS algorithm is iterated using the previous
estimates as starting values}

\item{tol}{Only used if iterate = TRUE and nIt = NULL. If the Euclidean distance
between the previous beta and current beta falls below tol, then the algorithm stops.}

\item{maxIt}{Only used it iterate = TRUE and nIt = NULL. Aborts iterative fitting
when number of iterations exceed maxIt}

\item{nIt}{Only used if iterate = TRUE. If this is specified, then tol is ignored
and the algorithm iterates nIt times. This option should not be used unless
the user has a specific reason to run the algorithm nIt times, e.g. for
replication purposes.}

\item{verbose}{If verbose = TRUE, then it prints the iteration process
(only relevant if iterate = TRUE).}

\item{formula}{an object of class "\link{Formula}"
(or one that can be coerced to that class):
a symbolic description of the model to be fitted.
The details of model specification are given under ‘Details’.}

\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame) containing the
variables in the model. If not found in data, the variables are taken from
environment(formula), typically the environment from which walsNB is called.}
}
\description{
Fits an NB2 regression model using the Weighted-Average Least Squares method
described in \insertCite{deluca2018glm;textual}{WALS}.

\strong{WARNING:} Interactions in formula do work work properly yet.
It is recommended to manually create the interactions beforehand and then
to insert them as 'linear terms' in the formula.
}
\details{
Formulas should always contain two parts, i.e. they should be of the form
"y ~ X11 + X12 | X21 + X22", where the variables before "|" are the focus
regressors (includes a constant by default) and the ones after "|" are the
auxiliary regressors.
}
\examples{
data("HMDA", package = "AER")
fitBinomial <- walsGLM(deny ~ pirat + hirat + lvrat + chist + mhist + phist |
                       selfemp + afam, data = HMDA, family = binomialWALS(),
                       prior = weibull())
summary(fitBinomial)

data("NMES1988", package = "AER")
fitPoisson <- walsGLM(emergency ~ health + chronic + age + gender |
                      I((age^2)/10) + married + region, data = NMES1988,
                      family = poissonWALS(), prior = laplace())
summary(fitPoisson)

}
\references{
\insertAllCited{}
}
