% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsGLM.R
\name{walsGLMfitIterate}
\alias{walsGLMfitIterate}
\title{Iteratively fitting walsGLM, internal function for walsGLM.formula and
walsGLM.matrix.}
\usage{
walsGLMfitIterate(
  y,
  X1,
  X2,
  family,
  na.action = NULL,
  weights = NULL,
  offset = NULL,
  prior = weibull(),
  controlGLMfit = list(),
  keepY = TRUE,
  keepX = FALSE,
  iterate = FALSE,
  tol = 1e-06,
  maxIt = 50,
  nIt = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{y}{Response as vector.}

\item{X1}{Design matrix for focus regressors. Usually includes a constant
(column full of 1s) and can be generated using model.matrix().}

\item{X2}{Design matrix for auxiliary regressors. Usually does not include
a constant column and can also be generated using model.matrix().}

\item{family}{Object of class \link{familyWALS}.}

\item{na.action}{Not implemented yet.}

\item{weights}{Not implemented yet.}

\item{offset}{Not implemented yet.}

\item{prior}{Object of class \code{"\link[WALS]{familyPrior}"}. For example
\code{\link[WALS]{weibull}} or \code{\link[WALS]{laplace}}.}

\item{controlGLMfit}{Controls estimation of starting values for one-step ML,
passed to \link[stats]{glm.fit}. See also \link[stats]{glm.control}.}

\item{keepY}{If \code{TRUE}, then output keeps response.}

\item{keepX}{If \code{TRUE}, then output keeps the design matrices.}

\item{iterate}{if TRUE then the WALS algorithm is iterated using the previous
estimates as starting values.}

\item{tol}{Only used if iterate = TRUE and nIt = NULL. If the Euclidean distance
between the previous beta and current beta falls below tol, then
the algorithm stops.}

\item{maxIt}{Only used if iterate = TRUE and nIt = NULL. Aborts iterative fitting
when number of iterations exceed maxIt.}

\item{nIt}{Only used if iterate = TRUE. If this is specified, then tol is ignored
and the algorithm iterates nIt times.}

\item{verbose}{If verbose = TRUE, then it prints the iteration process
(only relevant if iterate = TRUE).}

\item{...}{Arguments to be passed to the workhorse function walsGLMfit.}
}
\value{
A list containing all elements returned from \link[WALS]{walsGLMfit}
and additionally the following elements:
\item{y}{If \code{keepY = TRUE}, contains the response vector.}
\item{x}{list. If \code{keepX} is true, then it is a list with elements
\code{x1} and \code{x2} containing the design matrices of the focus and
auxiliary regressors, respectively.}
\item{weights}{returns the argument \code{weights}.}
\item{offset}{returns the argument \code{offset}.}
\item{converged}{Logical. Only relevant if \code{iterate = TRUE}. Equals
\code{TRUE} if iterative fitting converged, else \code{FALSE}. Is \code{NULL}
if \code{iterate = FALSE}.}
\item{it}{Number of iterations run in the iterative fitting algorithm.
\code{NULL} if \code{iterate = FALSE}.}
\item{deviance}{Deviance of the fitted regression model.}
\item{residuals}{Raw residuals, i.e. response - fitted mean.}
}
\description{
See description of \link[WALS]{walsGLM}.
}
\seealso{
\link{walsGLM}, \link{walsGLMfit}.
}
