% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsHurdle.R
\name{walsHurdle.fit}
\alias{walsHurdle.fit}
\title{Fitting function for WALS Hurdle models}
\usage{
\method{walsHurdle}{fit}(
  X1,
  X2,
  Z1,
  Z2,
  y,
  yBinary,
  betaStart1,
  betaStart2,
  gammaStart1,
  gammaStart2,
  family,
  familyZero,
  prior,
  priorZero,
  method,
  ...
)
}
\description{
Uses walsGLM.fit for the binary response (y > 0 vs y = 0)
and walsZerotrunc.fit for fitting a count data model on all y > 0
}
