% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsHurdle.R
\name{walsHurdle.formula}
\alias{walsHurdle.formula}
\title{Fits Hurlde count models with the Weighted-Average Least Squares method}
\usage{
\method{walsHurdle}{formula}(
  formula,
  data,
  subset,
  na.action,
  weights,
  offset,
  family,
  prior = weibull(),
  priorZero = weibull(),
  methodStart = "BFGS",
  controlStart = list(count = list(), zero = list()),
  model = TRUE,
  y = TRUE,
  x = FALSE,
  method = "original",
  ...
)
}
\description{
WARNING: Interactions in formula do work work properly yet, see bugs.txt
}
