% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsNB.R
\name{walsNB}
\alias{walsNB}
\alias{walsNB.formula}
\alias{walsNB.matrix}
\alias{walsNB.default}
\title{Weighted Average Least Squares Negative Binomial}
\usage{
walsNB(x, ...)

\method{walsNB}{formula}(
  formula,
  data,
  subset = NULL,
  na.action = NULL,
  weights = NULL,
  offset = NULL,
  link = "log",
  prior = weibull(),
  controlInitNB = controlNB(),
  model = TRUE,
  keepY = TRUE,
  keepX = FALSE,
  iterate = FALSE,
  tol = 1e-06,
  maxIt = 10000,
  nIt = NULL,
  verbose = FALSE,
  ...
)

\method{walsNB}{matrix}(
  X1,
  X2,
  y,
  link = "log",
  subset = NULL,
  na.action = NULL,
  weights = NULL,
  offset = NULL,
  prior = weibull(),
  controlInitNB = controlNB(),
  model = TRUE,
  keepY = TRUE,
  keepX = FALSE,
  iterate = FALSE,
  tol = 1e-06,
  maxIt = 10000,
  nIt = NULL,
  verbose = FALSE,
  ...
)

\method{walsNB}{default}(x, ...)
}
\arguments{
\item{...}{Arguments for workhorse \link[WALS]{walsNB.fit}.}

\item{formula}{an object of class "\link{Formula}"
(or one that can be coerced to that class):
a symbolic description of the model to be fitted.
The details of model specification are given under ‘Details’.}

\item{data}{an optional data frame, list or environment
(or object coercible by as.data.frame to a data frame) containing the
variables in the model. If not found in data, the variables are taken from
environment(formula), typically the environment from which walsNB is called.}

\item{subset}{an optional vector specifying a subset of observations to be
used in the fitting process.}

\item{na.action}{\strong{not implemented yet.}}

\item{weights}{\strong{not implemented yet.}}

\item{offset}{\strong{not implemented yet.}}

\item{link}{specifies the link function, currently only "log" is supported.}

\item{prior}{Object of class \code{familyPrior}. For example \link{weibull}
or \link{laplace}. Not tested with other priors.}

\item{controlInitNB}{Controls estimation of starting values for one-step ML,
see \link[WALS]{controlNB}.}

\item{model}{if \code{TRUE} (default), then the model.frame is stored in
the return.}

\item{keepY}{if \code{TRUE} (default), then the response is stored in
the return.}

\item{keepX}{if \code{TRUE}, then the model matrix is stored in the return.
the return.}

\item{iterate}{if TRUE then the WALS algorithm is iterated using the previous
estimates as starting values}

\item{tol}{Only used if iterate = TRUE and nIt = NULL. If the Euclidean distance
between the previous beta and current beta falls below tol and the absolute difference between
the previous and current rho falls below tol, then the algorithm stops.}

\item{maxIt}{Only used it iterate = TRUE and nIt = NULL. Aborts iterative fitting
when number of iterations exceed maxIt}

\item{nIt}{Only used if iterate = TRUE. If this is specified, then tol is ignored
and the algorithm iterates nIt times.}

\item{verbose}{If verbose = TRUE, then it prints the iteration process of
internal function \link[WALS]{walsNB.fitIterate} (only relevant if iterate = TRUE).}

\item{X1}{Design matrix for focus regressors. Usually includes a constant
(column full of 1's) and can be generated using model.matrix().}

\item{X2}{Design matrix for auxiliary regressors. Usually does not include
a constant column and can also be generated using model.matrix().}

\item{y}{Count response as vector}
}
\value{
For \code{walsNB.formula}, it returns an object of class
\code{walsNB} which inherits from \code{walsGLM} and \code{wals}.
See return of \link[WALS]{walsNB.fit} for more details.

For \code{walsNB.matrix}, it returns an object of class \code{walsNBmatrix},
which inherits from \code{walsNB}, \code{walsGLM} and \code{wals}.
See return of \link[WALS]{walsNB.fit} for more details.
}
\description{
Fits an NB2 regression model using the Weighted-Average Least Squares method
of \insertCite{huynhwalsnb;textual}{WALS}.

\strong{WARNING:} Interactions in formula do work work properly yet.
It is recommended to manually create the interactions beforehand and then
to insert them as 'linear terms' in the formula.

\code{walsNB.matrix()} uses prespecified design matrices X1 and X2 and
response vector y.
}
\details{
\code{walsNB.formula} uses formulas to specify the design matrix.

Formulas should always contain two parts, i.e. they should be of the form
"y ~ X11 + X12 | X21 + X22", where the variables before "|" are the focus
regressors (includes a constant by default) and the ones after "|" are the
auxiliary regressors.
}
\examples{
data("NMES1988", package = "AER")

fitWeibull <- walsNB(visits ~ health + chronic + age + gender | I((age^2)/10) +
                     married + region, data = NMES1988, prior = weibull())
summary(fitWeibull)

fitLaplace <- walsNB(visits ~ health + chronic + age + gender | I((age^2)/10) +
                     married + region, data = NMES1988, prior = laplace())
summary(fitLaplace)

}
\references{
\insertAllCited{}
}
