% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsNB.R
\name{walsNBfitIterate}
\alias{walsNBfitIterate}
\title{Iteratively fitting walsNB, internal function for walsNB.formula and
walsNB.vector.}
\usage{
walsNBfitIterate(
  y,
  X1,
  X2,
  link = "log",
  na.action = NULL,
  weights = NULL,
  offset = NULL,
  prior = weibull(),
  controlInitNB = controlNB(),
  keepY = TRUE,
  keepX = FALSE,
  iterate = FALSE,
  tol = 1e-06,
  maxIt = 10000,
  nIt = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{y}{Count response as vector}

\item{X1}{Design matrix for focus regressors. Usually includes a constant
(column full of 1's) and can be generated using model.matrix().}

\item{X2}{Design matrix for auxiliary regressors. Usually does not include
a constant column and can also be generated using model.matrix().}

\item{link}{specifies the link function, currently only "log" is supported.}

\item{na.action}{Not implemented yet.}

\item{weights}{Not implemented yet.}

\item{offset}{Not implemented yet.}

\item{prior}{Object of class \code{familyPrior}, e.g. \link[WALS]{weibull}.}

\item{controlInitNB}{Controls estimation of starting values for one-step ML,
see \link[WALS]{controlNB}.}

\item{keepY}{If \code{TRUE}, then output keeps response.}

\item{keepX}{If \code{TRUE}, then output keeps design matrix.}

\item{iterate}{if TRUE then the WALS algorithm is iterated using the previous
estimates as starting values}

\item{tol}{Only used if iterate = TRUE and nIt = NULL. If the Euclidean distance
between the previous beta and current beta falls below tol and the absolute difference between
the previous and current rho falls below tol, then the algorithm stops.}

\item{maxIt}{Only used if iterate = TRUE and nIt = NULL. Aborts iterative fitting
when number of iterations exceed maxIt.}

\item{nIt}{Only used if iterate = TRUE. If this is specified, then tol is ignored
and the algorithm iterates nIt times.}

\item{verbose}{If verbose = TRUE, then it prints the iteration process of
internal function walsNBfitIterate (only relevant if we iterate = TRUE).}

\item{...}{Arguments to be passed to the workhorse function walsNBfit which
actually fits the model.}
}
\description{
See description of \link{walsNB}.
}
