% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsZerotrunc.R
\name{walsZerotrunc}
\alias{walsZerotrunc}
\title{WALS for zerotruncated count data models}
\usage{
walsZerotrunc(x, ...)
}
\description{
Implements the WALS method for count data models (only for GLMs so far)
WARNING: Interactions in formula do work work properly yet, see bugs.txt
}
