% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{weibull}
\alias{weibull}
\title{Double (reflected) Weibull prior}
\usage{
weibull(q = 0.887630085544086, b = log(2))
}
\arguments{
\item{q}{\eqn{q} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution.}
}
\description{
Double (reflected) Weibull prior for Bayesian posterior mean estimation of
WALS estimators. Special case of the reflected generalized gamma distribution.
}
\details{
The default values for the parameters \code{q} and \code{b} are minimax regret
solutions. The double weibull prior is both neutral and robust.
See section 9 "Enter Bayes: Neutrality and Robustness" of
\insertCite{magnus2016wals;textual}{WALS} for details.
}
\references{
\insertAllCited{}
}
