% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/walsZerotrunc.R
\name{zerotruncBetaStart}
\alias{zerotruncBetaStart}
\title{Get starting values for WALS estimation}
\usage{
zerotruncBetaStart(X, Y, family, weights, offset, method, control)
}
\description{
Get starting values for WALS estimation with walsZerotrunc.fit via
maximizing likelihood with optim. Initial values are obtained via
maximizing the likelihood of a Poisson model via glm.fit.
}
