% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyWALS.R
\name{binomialWALS}
\alias{binomialWALS}
\title{Binomial family with additional functions}
\usage{
binomialWALS(link = "logit")
}
\arguments{
\item{link}{specifies link for \code{binomial} family.
See \link[stats]{family} for details.}
}
\description{
Extends \code{binomial} family from \code{stats} with transformations required
for \code{walsGLM}.
}
