% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{ddsubbotin}
\alias{ddsubbotin}
\title{Subbotin density}
\usage{
ddsubbotin(x, q, b, log = FALSE)
}
\arguments{
\item{q}{\eqn{q} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution (\link{ddgengamma}).}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution (\link{ddgengamma}).}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
Gives the (log-)density.
}
\description{
Wrapper around \link[VGAM]{dgengamma.stacy} of \link[VGAM]{VGAM} to use the
parametrization on pp. 131 of \insertCite{magnus2016wals;textual}{WALS}.
}
\references{
\insertAllCited{}
}
\seealso{
\link{subbotin}, \link{ddgengamma}.
}
