% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{ddweibull}
\alias{ddweibull}
\title{Internal function: double (reflected) Weibull density}
\usage{
ddweibull(x, q, b, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{q}{\eqn{q} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution. See below for details.}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution. See below for details.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
Gives the (log-)density.
}
\description{
Wrapper around \code{\link[stats]{dweibull}} to use the parametrization on
pp. 131 of \insertCite{magnus2016wals;textual}{WALS}.
}
\details{
The density function is
\deqn{\pi(x) = \frac{q c}{2} |x|^{q - 1} \exp(-c |x|^{q}).}
}
\references{
\insertAllCited{}
}
\seealso{
\link{weibull}, \link{dweibull}.
}
