% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{dlaplace}
\alias{dlaplace}
\title{Internal function: Laplace density}
\usage{
dlaplace(x, b, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution. See below for details.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
Gives the (log-)density.
}
\description{
Wrapper around \code{\link[WALS]{dsubbotin}} with fixed \code{q = 1}. Uses
the parametrization on pp. 131 of \insertCite{magnus2016wals;textual}{WALS}.
}
\details{
The density function is
\deqn{\pi(x) = \frac{c}{2} \exp(-c |x|).}
}
\references{
\insertAllCited{}
}
\seealso{
\link{laplace}, \link{dsubbotin}.
}
