% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPrior.R
\name{dsubbotin}
\alias{dsubbotin}
\title{Internal function: Subbotin density}
\usage{
dsubbotin(x, q, b, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{q}{\eqn{q} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution. See below for details.}

\item{b}{\eqn{c} in \insertCite{magnus2016wals;textual}{WALS}.
Parameter of reflected generalized gamma distribution. See below for details.}

\item{log}{logical; if TRUE, probabilities p are given as log(p).}
}
\value{
Gives the (log-)density.
}
\description{
Subbotin density, uses the parametrization on pp. 131 of
\insertCite{magnus2016wals;textual}{WALS}. Also called generalized normal
distribution.
}
\details{
The density function is
\deqn{\pi(x) = \frac{q c^{1/q}}{2 \Gamma(1/q)} \exp(-c |x|^{q}).}
}
\references{
\insertAllCited{}
}
\seealso{
\link{subbotin}.
}
