% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wals.R
\name{wals}
\alias{wals}
\alias{wals.matrix}
\alias{wals.default}
\title{Weighted Average Least Squares for linear regression models}
\usage{
wals(x, ...)

\method{wals}{matrix}(
  X1,
  X2,
  y,
  na.action,
  weights,
  offset,
  prior = weibull(),
  keepY = TRUE,
  keepX = FALSE,
  sigma = NULL,
  ...
)

\method{wals}{default}(x, ...)
}
\arguments{
\item{X1}{matrix of focus regressors.}

\item{X2}{matrix of auxiliary regressors.}

\item{y}{response vector.}
}
\description{
R port of MATLAB code wals.m (version 2.0, revision 18 december 2013)
by J.R. Magnus and G.De Luca, available from https://www.janmagnus.nl/items/WALS.pdf.
Calculates WALS estimates and variances when some regressors (X1) are present
in all models and model selection takes place over the rest (X2).
}
\examples{
X <- model.matrix(mpg ~ disp + hp + wt + vs + am + carb, data = mtcars)
X1 <- X[,c("(Intercept)", "disp", "hp", "wt")] # focus
X2 <- X[,c("vs", "am", "carb")] # auxiliary
y <- mtcars$mpg

wals(X1, X2, y, prior = weibull())

}
\references{
\insertRef{deluca2011stata}{WALS}

\insertRef{kumar2013normallocation}{WALS}

\insertRef{magnus2010growth}{WALS}

\insertRef{magnus2016wals}{WALS}
}
